feat(summary/repository):핵콕 저장소 기능 추가 및 파라미터 변경
- 핵콕 저장소 테스트를 위해 인터페이스에 저장소 초기화 기능을 추가함
- 핵콕 저장소 기능 중 수정과 삭제에 대상이 되는 객체의 아이디를 파라미터로 받로록 수정함

+++ src/main/java/com/cobin/hackok/domain/summary/repository/SummaryRepositoryImpl.java
@@ -1,10 +1,11 @@
 package com.cobin.hackok.domain.summary.repository;
 
 import com.cobin.hackok.domain.summary.dto.Summary;
+import org.springframework.stereotype.Repository;
 
 import java.util.HashMap;
 import java.util.Map;
-
+@Repository
 public class SummaryRepositoryImpl implements SummaryRepository{
 
     // 데이터 베이스 연결 전 사용할 임시 저장소
@@ -25,15 +26,23 @@ public class SummaryRepositoryImpl implements SummaryRepository{
 
     // 3. 수정
     @Override
-    public Summary update(Summary summary) {
-        repo.replace(summary.getId(), repo.get(summary.getId()), summary);
+    public Summary update(Long oldId, Summary summary) {
+        repo.replace(oldId, repo.get(summary.getId()), summary);
         return repo.get(summary.getId());
     }
 
     // 4. 삭제
     @Override
     public Boolean delete(Summary summary) {
-        repo.remove(summary.getId());
-        return true;
+        if(read(summary) != null) {
+            repo.remove(summary.getId());
+            return  true;
+        }
+        return false;
+    }
+
+    @Override
+    public void clearStore() {
+        repo.clear();
     }
 }
